src/main/java/easy/TwoSum.java
package easy;

public class TwoSum {
    public int[] twoSum(int[] nums, int target) {
        int[] result = new int[2];
        boolean isFind = false;
        for (int i = 0; i < nums.length - 1; i++) {
            for (int j = i + 1; j < nums.length; j++) {
                if (nums[i] + nums[j] == target) {
                    result[0] = i;
                    result[1] = j;
                    isFind = true;
                    break;
                }
            }
            if (isFind) {
                break;
            }
        }
        return result;
    }
}
